+++ facebook_sublet_group_bot/util.py
@@ -1,8 +1,23 @@
 import sys
 import subprocess
 
+'''
+Project:    RASCAL: Robotic Autonomous Space Cadet Adminstration Lackey
+File:       util.py
+Author:     Whitney King (@WhitneyOnTheWeb)
+Fork:       Henri Sweers FB Modbot (2014 / Py2)
 
-# Color class, used for colors in terminal
+Log:
+    - 26/06/2019 (v1.0):
+      - Forked repro of old bot used for automoding roomate search FB groups
+      - Restructure from Python 2 -> 3.7
+      - Add scaffolding for Windows/Mac features
+      - Redesign for generalized group administration features
+      - Design customized TOCA group administrative features
+'''
+
+# Color Class:
+# Terminal Display Colors
 class Color:
     PURPLE = '\033[95m'
     CYAN = '\033[96m'
@@ -16,15 +31,17 @@ class Color:
     END = '\033[0m'
 
 
-# Log method. If there's a color argument, it'll stick that in first
+# Log Method:
+# Checks for Color, then adds log
 def log(message, *colorargs):
     if len(colorargs) > 0:
-        print colorargs[0] + message + Color.END
+        print(colorargs[0] + message + Color.END)
     else:
-        print message
+        print (message)
 
 
-# Nifty method for sending notifications on my mac when it's done
+# Notify Method:
+# Currently just notifies on Mac, needs update for Windows/Linux
 def notify_mac():
     if sys.platform == "darwin":
         try:
@@ -32,11 +49,12 @@ def notify_mac():
                 ["terminal-notifier", "-message", "Tests done", "-title",
                  "FB_Bot", "-sound", "default"])
         except OSError:
-            print "If you have terminal-notifier, this would be a notification"
+            print("If you have terminal-notifier, this would be a notification")
 
 
-# Reads in multi-line strings based on a certain "newline" type
-# Borrowed from here: http://stackoverflow.com/a/16260159/3034339
+# Read_Lines Method:
+# Reads in multi-line strings based on "newline" type
+# http://stackoverflow.com/a/16260159/3034339
 def read_lines(f, newline):
     buf = ""
     while True: