+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-FB Mod Bot Changelog
-====================
+RASCAL Changelog
+================
 
-### 02/08/14
-* Created changelog
-* Added `[PARKING]` tag checking
+### 02/08/14 (ZS)
+
+- Created changelog
+- Added `[PARKING]` tag checking
+
+### 26/06/2019 (WK)
+
+- 26/06/2019 (v1.0):
+    - Forked repro of old bot used for automodding roommate search FB groups
+    - Restructure from Python 2 -> 3.7
+    - Add scaffolding for Windows/Mac features
+    - Redesign for generalized group administration features
+    - Design customized TOCA group administrative features